requirements.txt
streamlit
google-genai
mutagen
tenacity
diff-match-patch
